zayavki_service/requirements.txt
django>=5.2
 djangorestframework
drf-yasg
psycopg2-binary 